# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "myrti",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/files/original/{id}": {
      "get": {
        "tags": [
          "crate::routes::file"
        ],
        "operationId": "getOriginalFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "FileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/repr/{fileId}/{reprId}": {
      "get": {
        "tags": [
          "crate::routes::file"
        ],
        "operationId": "getImageAssetRepresentation",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "description": "FileId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FileId"
            }
          },
          {
            "name": "reprId",
            "in": "path",
            "description": "ImageRepresentationId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ImageRepresentationId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "File or Representation not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/thumbnail/{id}/{size}/{format}": {
      "get": {
        "tags": [
          "crate::routes::file"
        ],
        "operationId": "getThumbnail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "FileId to get thumbnail for",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FileId"
            }
          },
          {
            "name": "size",
            "in": "path",
            "description": "Thumbnail size",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ThumbnailSize"
            }
          },
          {
            "name": "format",
            "in": "path",
            "description": "Image format for thumbnail",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ThumbnailFormat"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{id}/details": {
      "get": {
        "tags": [
          "crate::routes::file"
        ],
        "operationId": "getFileDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "FileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetailsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Asset not found"
          }
        }
      }
    },
    "/api/files/{id}/transform": {
      "post": {
        "tags": [
          "crate::routes::file"
        ],
        "operationId": "setAssetTransformCorrection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "FileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAssetRotationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetWithSpe"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Asset": {
        "type": "object",
        "required": [
          "assetId",
          "takenDate",
          "repFile"
        ],
        "properties": {
          "assetId": {
            "$ref": "#/components/schemas/AssetId"
          },
          "repFile": {
            "$ref": "#/components/schemas/AssetFile"
          },
          "takenDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AssetDetailsResponse": {
        "type": "object",
        "required": [
          "exiftoolOutput"
        ],
        "properties": {
          "exiftoolOutput": {}
        }
      },
      "AssetFile": {
        "type": "object",
        "required": [
          "fileId",
          "assetRootId",
          "pathInRoot",
          "width",
          "height",
          "addedAt",
          "rotationCorrection",
          "mirrorCorrection",
          "mimeType"
        ],
        "properties": {
          "addedAt": {
            "type": "string",
            "format": "date-time"
          },
          "assetRootId": {
            "$ref": "#/components/schemas/AssetRootDirId"
          },
          "fileId": {
            "$ref": "#/components/schemas/FileId"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "mimeType": {
            "type": "string"
          },
          "mirrorCorrection": {
            "$ref": "#/components/schemas/MirrorCorrection"
          },
          "pathInRoot": {
            "type": "string"
          },
          "rotationCorrection": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AssetId": {
        "type": "string"
      },
      "AssetRootDirId": {
        "type": "string"
      },
...
      "AssetSpe": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Image"
              },
              {
                "type": "object",
                "required": [
                  "assetType"
                ],
                "properties": {
                  "assetType": {
                    "type": "string",
                    "enum": [
                      "image"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Video"
              },
              {
                "type": "object",
                "required": [
                  "assetType"
                ],
                "properties": {
                  "assetType": {
                    "type": "string",
                    "enum": [
                      "video"
                    ]
                  }
                }
              }
            ]
          }
        ],
        "discriminator": {
          "propertyName": "assetType"
        }
      },
...
      "AssetWithSpe": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Asset"
          },
          {
            "$ref": "#/components/schemas/AssetSpe"
          },
          {
            "type": "object"
          }
        ]
      },
...
      "FileId": {
        "type": "string"
      },
...
      "Image": {
        "type": "object",
        "required": [
          "representations"
        ],
        "properties": {
          "representations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageRepresentation"
            }
          }
        }
      },
      "ImageRepresentation": {
        "type": "object",
        "required": [
          "id",
          "format",
          "width",
          "height",
          "size"
        ],
        "properties": {
          "format": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ImageRepresentationId": {
        "type": "string"
      },
      "MirrorCorrection": {
        "type": "string",
        "enum": [
          "None",
          "Horizontal",
          "Vertical"
        ]
      },
...
      "SetAssetRotationRequest": {
        "type": "object",
        "properties": {
          "mirror": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MirrorCorrection"
              }
            ],
            "nullable": true
          },
          "rotation": {
            "type": "integer",
            "format": "int32",
            "description": "One of 0, 90, 180, 270",
            "nullable": true
          }
        }
      },
...
      "ThumbnailFormat": {
        "type": "string",
        "enum": [
          "avif",
          "webp"
        ]
      },
      "ThumbnailSize": {
        "type": "string",
        "enum": [
          "small",
          "large"
        ]
      },
...
      "Video": {
        "type": "object",
        "required": [
          "hasDash"
        ],
        "properties": {
          "hasDash": {
            "type": "boolean"
          }
        }
      }
    }
  },
...
}